feat: optimistic member updates + decompose QuickAddAgentPopover
- Add optimistic onMutate to useAttachManagedAgentToChannelMutation
  (instant member insertion, rollback on error)
- Add onSuccess cache-set to create/batch mutations (member appears
  as soon as server confirms, before invalidation round-trip)
- Decompose QuickAddAgentPopover (816→271 lines) into:
  - useQuickAddAgentItems.ts (data fetching + item list building)
  - useQuickAddAgentActions.ts (handlers + multi-select state)
  - QuickAddAgentItemRow.tsx (memoized item row component)
- Fix ARIA: remove conflicting role="menu" on container, add
  role="group" with aria-label for keyboard navigation wrapper
- Fix dead team toggles: filter out teams whose members are all
  already in the channel
